:building_construction: Create a SearchMessage structure
