+++ src/commands/basic/errorDetails.js
@@ -1,5 +1,5 @@
 import Command from '../../structures/Command';
-import { db } from '../../main';
+import searchMessage from '../../structures/SearchMessage';
 
 class ErrorDetails extends Command {
   constructor() {
@@ -9,40 +9,8 @@ class ErrorDetails extends Command {
     this.examples = ["errordetail Can't compare 'if arg 1' with a text"];
   }
 
-  async execute(_client, message, args) {
-    const arg = args.join(' ').toLowerCase();
-    const messages = await db.messages.find({ type: 'error' }).catch(console.error);
-
-    if (args.length === 0) {
-      message.channel.sendError(this.config.noArg.replace('%s', `\`${messages.map(msg => msg.title).join('`, `')}\``), message.member);
-      return;
-    }
-
-    for (const msg of messages) {
-      for (const aliase of msg.aliases) {
-        if (arg.includes(aliase) || arg.includes(msg.title)) {
-          const msgContent = await message.channel.send(msg.content);
-          msgContent.react('🗑️');
-          const removeCollector = msgContent
-            .createReactionCollector((reaction, user) => reaction.emoji.name === '🗑️' && user.id === message.author.id)
-            .once('collect', () => {
-              removeCollector.stop();
-              msgContent.delete();
-              message.delete();
-            });
-          return;
-        }
-      }
-    }
-    const msg = await message.channel.send(this.config.invalidMessage);
-    msg.react('🗑️');
-    const removeCollector = msg
-      .createReactionCollector((reaction, user) => reaction.emoji.name === '🗑️' && user.id === message.author.id)
-      .once('collect', () => {
-        removeCollector.stop();
-        msg.delete();
-        message.delete();
-      });
+  async execute(client, message, args) {
+    await searchMessage(client, message, args, 'error');
   }
 }
